Move widgets to seperate package
