+++ src/main/java/org/terasology/tutorialnui/widgets/UIButtonScreen.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.tutorialnui;
+package org.terasology.tutorialnui.widgets;
 
 import org.terasology.rendering.nui.CoreScreenLayer;
 
-public class UIImageScreen extends CoreScreenLayer {
-
+public class UIButtonScreen extends CoreScreenLayer {
     @Override
     public void initialise() {
