+++ src/main/java/org/terasology/tutorialnui/widgets/UICheckboxScreen.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.tutorialnui;
+package org.terasology.tutorialnui.widgets;
 
 import org.terasology.rendering.nui.CoreScreenLayer;
 
-public class UIBoxScreen extends CoreScreenLayer {
+public class UICheckboxScreen extends CoreScreenLayer {
     @Override
-    public void initialise() {
+    public void initialise(){
 
     }
 }
