+++ src/main/java/org/terasology/tutorialnui/widgets/UIIconBarScreen.java
@@ -13,11 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.tutorialnui;
+package org.terasology.tutorialnui.widgets;
 
 import org.terasology.rendering.nui.CoreScreenLayer;
 
-public class UIButtonScreen extends CoreScreenLayer {
+public class UIIconBarScreen extends CoreScreenLayer {
+
     @Override
     public void initialise() {
