+++ src/main/java/org/terasology/tutorialnui/widgets/UIImageScreen.java
@@ -13,22 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.tutorialnui;
+package org.terasology.tutorialnui.widgets;
 
 import org.terasology.rendering.nui.CoreScreenLayer;
-import org.terasology.rendering.nui.widgets.UIDropdownScrollable;
 
-import java.util.ArrayList;
-import java.util.List;
-
-public class UIDropdownScrollableScreen extends CoreScreenLayer {
+public class UIImageScreen extends CoreScreenLayer {
 
     @Override
     public void initialise() {
-        UIDropdownScrollable dropdown = find("Dropdown", UIDropdownScrollable.class);
-        List<String> options = new ArrayList<>();
-        for (int i = 1; i <= 10; i++)
-            options.add("Option " + i);
-        dropdown.setOptions(options);
+
     }
 }
