+++ src/main/java/org/terasology/tutorialnui/widgets/UILabelScreen.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.tutorialnui;
+package org.terasology.tutorialnui.widgets;
 
 import org.terasology.rendering.nui.CoreScreenLayer;
 
-public class UIScrollbarScreen extends CoreScreenLayer {
+public class UILabelScreen extends CoreScreenLayer {
 
     @Override
     public void initialise() {
