+++ src/main/java/org/terasology/tutorialnui/widgets/UIScrollingTextScreen.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.tutorialnui;
-
+package org.terasology.tutorialnui.widgets;
 
 import org.terasology.rendering.nui.CoreScreenLayer;
 
-public class UIDoubleSliderScreen extends CoreScreenLayer {
+public class UIScrollingTextScreen extends CoreScreenLayer {
+
     @Override
     public void initialise() {
